Move symlinks step to a separate step so they could be used safely without having an order of operations problem.
