+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Next version
 
-- Put your changes here...
+- Moved symlink creation to a separate step and made it the first thing that is done on app startup
 
 ## 0.19.3
