+++ lib/mapRoutes.js
@@ -15,7 +15,6 @@ module.exports = app => {
   const fsr = require('./tools/fsr')(app)
   const params = app.get('params')
   const prefix = params.routePrefix
-  const publicDir = params.publicFolder
   let controllerFiles
 
   // ensure 404 page exists
@@ -74,35 +73,6 @@ module.exports = app => {
       res.on('error', function () { params.onReqAfterRoute(req, res) })
       next()
     })
-  }
-
-  // generate public and statics directories
-  fsr.ensureDirSync(publicDir)
-  fsr.ensureDirSync(params.staticsRoot)
-
-  // process symlinks
-  if (params.generateFolderStructure) {
-    for (const symlink of params.symlinks) {
-      // append appDir to each path that is relative
-      const source = path.isAbsolute(symlink.source) ? symlink.source : path.join(app.get('appDir'), symlink.source)
-      const dest = path.isAbsolute(symlink.dest) ? symlink.dest : path.join(app.get('appDir'), symlink.dest)
-
-      // first ensure the source exists
-      if (fsr.fileExists(source)) {
-        // then check if the destination already exists
-        if (fsr.fileExists(dest)) {
-          // then check if the destination is a symlink
-          if (!fs.lstatSync(dest).isSymbolicLink()) {
-            logger.error(`Symlink destination "${dest}" is already a file that exists. Skipping symlink creation.`)
-          }
-        } else {
-          fs.ensureSymlinkSync(source, dest)
-          logger.info('📁', `${appName} making new symlink `.cyan + `${dest}`.yellow + (' pointing to ').cyan + `${source}`.yellow)
-        }
-      } else {
-        logger.error(`Symlink source "${source}" does not exist. Skipping symlink creation.`)
-      }
-    }
   }
 
   // generate mvc directories
